backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pyserial==3.5
requests==2.31.0
python-dotenv==1.0.1
websockets>=13.0
supabase
